Split new proxy implementations for HTTP and gRPC into separate packages
