--- trac-services/trac-gateway/src/main/java/com/accenture/trac/gateway/routing/Http1Client.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2021 Accenture Global Solutions Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.accenture.trac.gateway.routing;
-
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.channel.SimpleChannelInboundHandler;
-import io.netty.handler.codec.http.HttpObject;
-
-
-public class Http1Client extends SimpleChannelInboundHandler<HttpObject> {
-
-    @Override
-    protected void channelRead0(ChannelHandlerContext ctx, HttpObject msg) throws Exception {
-
-    }
-}
